requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-google-genai
google-generativeai
pandas
python-dotenv